+++ README.md
@@ -19,6 +19,8 @@ See the [docs](docs) for some detail on design choices.
 
 ## Design
 
+### Containers
+
 Fluxqueue builds three primary containers:
 
  - `ghcr.io/converged-computing/fluxqueue`: contains the webhook and operator with a flux queue for pods and groups that interacts with fluxion
@@ -26,6 +28,10 @@ Fluxqueue builds three primary containers:
  - `ghcr.io/converged-computing/fluxqueue-postgres`: holds the worker queue and provisional queue tables
 
 And we use `ghcr.io/converged-computing/fluxion` for the fluxion service.
+
+### Choices
+
+- **Duration of job comes from Kubernetes** Right now, we don't allow a special or different duration to be given to Fluxion. Any duration or deletion needs to come from Kubernetes first, by way of an object deletion. Otherwise we would need to orchestrate deletion from the cluster and Fluxion, and it's easier to ask the user to delete with a job duration or other mechanism.
 
 ## Deploy
 
@@ -212,14 +218,12 @@ SELECT group_name, group_size from pods_provisional;
 
 ### TODO
 
-- [ ] node separator of `__` is ok?
-- [ ] another mechanism to move schedule queue aside from new submission
-- [ ] How does fluxion work with a duration - does it remove scheduled stuff at that timestamp?
-  - Need to figure out way to coordinate fluxion time ending with cancel of pods, and cancel of pods with cancel to fluxion.
-  - The latter is OK if there is some delay, because fluxion wouldn't reassign to nodes it thinks are being used.
-  - There is mostly just the wasted resources time in that. 
-  - But we can't cancel to fluxion if pods still running on nodes - something else could be scheduled.
-- [ ] I already forgot how reservations work!
+- [ ] For cancel, we would issue a cancel for every pod associated with a job. How can we avoid that (or is that OK?)
+- [ ] we will eventually need another mechanism to move schedule queue aside from new submission
+- [ ] River is task based, so the entire concept of a loop is not relevant. We are asking flux asyncronously. The schedule loop, in that it kicks of different river jobs, isn't actually a loop - they run in sync, potentially. 
+   - So the reservation strategy doesn't make sense because there is no "end" point.
+   - Currently I'm just going to set the reservationDepth to 0
+- Reservations notes:
  - We have a queue that sends jobs to the scheduler
  - The scheduler does everything that it can, and then it starts reserving things
  - Reservations block off resources and give an estimated start time.
@@ -227,10 +231,6 @@ SELECT group_name, group_size from pods_provisional;
  - If reservation set to true, it will keep looking for earliest time in future.
  - Reservations are saving state of those jobs for the scheduler JUST during the loop
  - The reservations are cleared out after the loop.
- - Make easy / basic backfill reservations on by default
- - [ ] The deletion of objects should always trigger fluxion cancel
- - [ ] objects that don't have end time need to submit a delete job
- - [ ] cleanup job will issue the cancel to fluxion, respond to event
 - [ ] scheduleAt can be used to AskFlux in the future
 - [ ] Nodes that are currently assigned need to be taken into account
    - Right now they aren't included in resources, but instead should be "given" to Fluxion.
@@ -241,7 +241,6 @@ SELECT group_name, group_size from pods_provisional;
    - make up job ids
    - take system / operator pods that are running, convert into JGF, and then figure out which resources each are running on. 
    - at end of loop need to go through jobs, for those that aren't scheduled cancel.
-- [ ] deletion of job or pod needs to trigger deletion / cancel
 - [ ] kubectl plugin to get fluxion state?
 
 ## License
