+++ cmd/manager/main.go
@@ -28,7 +28,6 @@ import (
 	"k8s.io/client-go/rest"
 
 	fluxion "github.com/converged-computing/fluxion/pkg/client"
-	"github.com/riverqueue/river"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
@@ -243,23 +242,6 @@ func main() {
 		setupLog.Error(err, "problem running manager")
 		os.Exit(1)
 	}
-
-	// TODO do we actually need this?
-	waitForJob := func(subscribeChan <-chan *river.Event) {
-		for {
-			select {
-			case event := <-subscribeChan:
-				if event == nil {
-					setupLog.Info("Channel is closed")
-					return
-				}
-				setupLog.Info("Job Event Received", "Kind", event.Job.Kind)
-			}
-		}
-	}
-	defer queue.EventChannel.Function()
-	waitForJob(queue.EventChannel.Channel)
-
 	err = queue.Stop(ctx)
 	if err != nil {
 		setupLog.Error(err, "Failed to stop FluxQueue")
