+++ pkg/fluxqueue/events.go
@@ -1,85 +1,29 @@
 package fluxqueue
 
 import (
-	corev1 "k8s.io/api/core/v1"
-	klog "k8s.io/klog/v2"
+	"context"
+	"os"
+
+	"github.com/converged-computing/fluxqueue/pkg/fluxqueue/queries"
+	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-// UpdatePodEvent is called on an update, and the old and new object are presented
-func (q *Queue) UpdatePodEvent(oldObj, newObj interface{}) {
+// deleteFromPending ensures that the job is removed from pending
+// This happens on a job that is impossible to schedule, or one that is successfully running
+func deleteFromPending(namespace, name string) error {
 
-	pod := oldObj.(*corev1.Pod)
-	newPod := newObj.(*corev1.Pod)
-
-	// a pod is updated, get the group. TODO: how to handle change in group name?
-	// groupName := groups.GetPodGroupName(oldPod)
-	switch pod.Status.Phase {
-	case corev1.PodPending:
-		klog.Infof("Received update event 'Pending' to '%s' for pod %s/%s", newPod.Status.Phase, pod.Namespace, pod.Name)
-	case corev1.PodRunning:
-		klog.Infof("Received update event 'Running' to '%s' for pod %s/%s", newPod.Status.Phase, pod.Namespace, pod.Name)
-	case corev1.PodSucceeded:
-		klog.Infof("Received update event 'Succeeded' to '%s' for pod %s/%s", newPod.Status.Phase, pod.Namespace, pod.Name)
-	case corev1.PodFailed:
-		klog.Infof("Received update event 'Failed' to '%s' for pod %s/%s", newPod.Status.Phase, pod.Namespace, pod.Name)
-	case corev1.PodUnknown:
-		klog.Infof("Received update event 'Unknown' to '%s' for pod %s/%s", newPod.Status.Phase, pod.Namespace, pod.Name)
-	default:
-		klog.Infof("Received unknown update event %s for pod %s/%s", pod.Status.Phase, pod.Namespace, pod.Name)
+	pool, err := pgxpool.New(context.Background(), os.Getenv("DATABASE_URL"))
+	if err != nil {
+		qlog.Info("Issue creating new pool", "Error", err.Error())
+		return err
 	}
-}
+	defer pool.Close()
 
-// DeletePodEventhandles the delete event handler
-// We don't need to worry about calling cancel to fluxion if the fluxid is already cleaned up
-// It has a boolean that won't return an error if the job does not exist.
-func (q *Queue) DeletePodEvent(podObj interface{}) {
-	pod := podObj.(*corev1.Pod)
-
-	switch pod.Status.Phase {
-	case corev1.PodPending:
-		klog.Infof("Received delete event 'Pending' for pod %s/%s", pod.Namespace, pod.Name)
-	case corev1.PodRunning:
-		klog.Infof("Received delete event 'Running' for pod %s/%s", pod.Namespace, pod.Name)
-	case corev1.PodSucceeded:
-		klog.Infof("Received delete event 'Succeeded' for pod %s/%s", pod.Namespace, pod.Name)
-	case corev1.PodFailed:
-		klog.Infof("Received delete event 'Failed' for pod %s/%s", pod.Namespace, pod.Name)
-	case corev1.PodUnknown:
-		klog.Infof("Received delete event 'Unknown' for pod %s/%s", pod.Namespace, pod.Name)
-	default:
-		klog.Infof("Received unknown update event %s for pod %s/%s", pod.Status.Phase, pod.Namespace, pod.Name)
+	_, err = pool.Exec(context.Background(), queries.DeleteFromPendingQuery, name, namespace)
+	if err != nil {
+		qlog.Info("Error deleting job from pending queue", "Namespace", namespace, "Name", name)
+		return err
 	}
-	// Get the fluxid from the database, and issue cleanup for the group:
-	// - deletes fluxID if it exists
-	// - cleans up Kubernetes objects up to parent with "true"
-	// - cleans up job in pending table
-	//podspec, err := json.Marshal(pod)
-	//if err != nil {
-	//	klog.Errorf("Issue marshalling podspec for Pod %s/%s", pod.Namespace, pod.Name)
-	//}
-	//groupName := groups.GetPodGroupName(pod)
-
-	// Since this is a termination event (meaning a single pod has terminated)
-	// we only want to cancel the fluxion job if ALL pods in the group are done.
-	// We don't want to delete the Kubernetes objects - this should happen on its
-	// own, never (if no timeout) or with a cancel job if a duration is set
-	//	pods, err := q.GetGroupPods(pod.Namespace, "")
-	//	if err != nil {
-	//		klog.Errorf("Issue getting group pods for %s", "")
-	//	}
-	//fluxID, err := q.GetFluxID(pod.Namespace, "")
-
-	// Determine finished status (delete via fluxion flux id ONLY if all are finished)
-	//	finished := true
-	//	for _, pod := range pods {
-	//		isFinished := podutil.IsPodPhaseTerminal(pod.Status.Phase)
-	//		if !isFinished {
-	//			finished = false
-	//			break
-	//		}
-	//	}
-	//	if !finished {
-	//		fluxID = -1
-	//	}
-	//err = workers.Cleanup(q.Context, string(podspec), fluxID, false, "")
+	qlog.Info("Removed job from pending", "Namespace", namespace, "Name", name)
+	return nil
 }
